Moved serializers to one file
