--- swarmdjango/core/serializers/SponsorSerializer.py
@@ -1,8 +1,0 @@
-from rest_framework import serializers
-from core.models import Sponsor
-
-
-class SponsorSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Sponsor
-        fields = ('companyName', 'page')
