--- swarmdjango/core/serializers/UserSerializer.py
@@ -1,8 +1,0 @@
-from rest_framework import serializers
-from core.models import User
-
-
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = ('username', 'password', 'email', 'firstName', 'lastName')
